13 Juni 2024

+++ example-app/app/Http/Controllers/MahasiswaController.php
@@ -121,6 +121,9 @@ class MahasiswaController extends Controller
      */
     public function destroy(Mahasiswa $mahasiswa)
     {
+        if(auth()->user()->cannot('delete', $mahasiswa)){
+            abort(403);
+        }
         //dd($mahasiswa);
         File::delete('foto/'. $mahasiswa['url_foto']);
         $mahasiswa->delete();
